refact: Avatar 컴포넌트 스타일 코드 분리
